comments and typo fixes to README.Rmd
Minor nitpick: heatmaps could be much smaller, fields could be square and you could add the group labels to the heatmaps so we can see that groups cluster etc.
Also I think there is a way to plot numbers directly onto each field, maybe that way you can plot distance + uncertainty in one heatmap. Easier to read than 2 heatmaps imo.

"If we were to call dist_approx directly with the params object this could bias our result, because the missing values would be inferred using the group structure. But usually this is exactly what we want to check in the first place. Instead we transform the parameter object and act as if every sample belonged to the same condition."
Maybe add some flag to dist_approx, like unbiased=T that ignores the group structure? Maybe set this to default, just to be on the safe side?

"We know that 10% of the data was changed(...)"
At this point the reader doesn't know this yet, I think. Maybe mention this when you generate the simulated data.

"(...) to look for both conditions how many values were actually observed and separate the data by this information (...)"
I know what you mean cause of your talk, but I think this could be formulated more clearly, maybe give some example like "3-2 means ...". Or maybe it's enough to add some facet label to the plot.